+++ wombat/windfarm/system/cable.py
@@ -1,8 +1,9 @@
 """"Defines the Cable class and cable simulations."""
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Generator
+from typing import TYPE_CHECKING
 from itertools import zip_longest
+from collections.abc import Generator
 
 import numpy as np
 import simpy